+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/components/sharedelement/RNSharedElementTransitionManager.java
@@ -4,12 +4,14 @@ import java.util.Map;
 
 import android.view.View;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.facebook.react.common.MapBuilder;
 import com.facebook.react.uimanager.annotations.ReactProp;
 import com.facebook.react.uimanager.ThemedReactContext;
 import com.facebook.react.uimanager.SimpleViewManager;
 import com.facebook.react.bridge.ReadableMap;
-import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactApplicationContext;
 
 public class RNSharedElementTransitionManager extends SimpleViewManager<RNSharedElementTransition> {
@@ -19,14 +21,16 @@ public class RNSharedElementTransitionManager extends SimpleViewManager<RNShared
     super();
   }
 
+  @NonNull
   @Override
   public String getName() {
     return REACT_CLASS;
   }
 
+  @Nullable
   @Override
-  public Map getExportedCustomBubblingEventTypeConstants() {
-    return MapBuilder.builder()
+  public Map<String, Object> getExportedCustomBubblingEventTypeConstants() {
+    return MapBuilder.<String, Object>builder()
             .put(
                     "onMeasureNode",
                     MapBuilder.of(
@@ -35,14 +39,15 @@ public class RNSharedElementTransitionManager extends SimpleViewManager<RNShared
             .build();
   }
 
+  @NonNull
   @Override
   public RNSharedElementTransition createViewInstance(ThemedReactContext reactContext) {
-    RNSharedElementModule module = (RNSharedElementModule) reactContext.getNativeModule(RNSharedElementModule.class);
+    RNSharedElementModule module = reactContext.getNativeModule(RNSharedElementModule.class);
     return new RNSharedElementTransition(reactContext, module.getNodeManager());
   }
 
   @Override
-  public void onDropViewInstance(RNSharedElementTransition view) {
+  public void onDropViewInstance(@NonNull RNSharedElementTransition view) {
     super.onDropViewInstance(view);
     view.releaseData();
   }