+++ ios/Exponent/Versioned/Core/Api/Components/SharedElement/RNSharedElementTransitionItem.h
@@ -13,13 +13,13 @@
 @property (nonatomic, readonly) RNSharedElementNodeManager* nodeManager;
 @property (nonatomic, readonly) BOOL isAncestor;
 @property (nonatomic, readonly) NSString* name;
-@property (nonatomic, assign) RNSharedElementNode* node;
+@property (nonatomic, strong) RNSharedElementNode* node;
 @property (nonatomic, assign) BOOL hidden;
 @property (nonatomic, assign) BOOL needsLayout;
 @property (nonatomic, assign) BOOL needsContent;
 @property (nonatomic, assign) BOOL hasCalledOnMeasure;
-@property (nonatomic, assign) RNSharedElementStyle* style;
-@property (nonatomic, assign) RNSharedElementContent* content;
+@property (nonatomic, strong) RNSharedElementStyle* style;
+@property (nonatomic, strong) RNSharedElementContent* content;
 
 - (instancetype)initWithNodeManager:(RNSharedElementNodeManager*)nodeManager name:(NSString*)name isAncestor:(BOOL)isAncestor;